Update fr.json
Translation of the missing items. 

Also, in this submitted translation, I proposed to rename "backlinks" into "rétroliens" instead of "liens retour" (as I did this summer), the reason is : the translation is still correct, its shorter and also its much more easy to pronounce. I experienced it when I was giving workshops on Obsidian to my students.